MockDatabase renamed to decoratorBase for refactoring
